requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0
groq>=0.4.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
moviepy==1.0.3
librosa==0.10.1
soundfile==0.12.1
reportlab==4.0.4
matplotlib==3.7.2
seaborn==0.12.2
openai-whisper==20231117
SpeechRecognition==3.10.0
pyaudio==0.2.11
